requirements.txt
# Core dependencies
requests
flask
flask-cors
python-dotenv


# Optional — only needed if you use Gemini or another LLM
google-generativeai


